requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.0
lxml==4.5.0
numpy==1.18.1
pandas==1.0.3
requests==2.22.0
psycopg2==2.8.4
